feat: Improve with CommandRunner
