+++ src/adaptor/proxy/command/schema.test.ts
