+++ src/service/command/guild-info.test.ts
@@ -3,7 +3,7 @@ import { afterEach, describe, expect, it, vi } from 'vitest';
 
 import type { Snowflake } from '../../model/id.js';
 import { createMockMessage } from './command-message.js';
-import { parseStringsOrThrow } from '../../adaptor/proxy/middleware/message-convert/schema.js';
+import { parseStringsOrThrow } from '../../adaptor/proxy/command/schema.js';
 
 describe('GuildInfo', () => {
   afterEach(() => {
@@ -39,7 +39,6 @@ describe('GuildInfo', () => {
     const fn = vi.fn();
 
     await guildInfo.on(
-      'CREATE',
       createMockMessage(
         parseStringsOrThrow(['guildinfo'], guildInfo.schema),
         fn
@@ -130,7 +129,6 @@ describe('GuildInfo', () => {
     const fn = vi.fn();
 
     await guildInfo.on(
-      'CREATE',
       createMockMessage(
         parseStringsOrThrow(['guildinfo'], guildInfo.schema),
         fn
