+++ src/service/command/guild-info.ts
@@ -4,7 +4,6 @@ import type {
   HelpInfo
 } from './command-message.js';
 
-import type { MessageEvent } from '../../runner/message.js';
 import type { Snowflake } from '../../model/id.js';
 import { createTimestamp } from '../../model/create-timestamp.js';
 
@@ -84,14 +83,7 @@ export class GuildInfo implements CommandResponder<typeof SCHEMA> {
 
   constructor(private readonly repo: GuildStatsRepository) {}
 
-  async on(
-    event: MessageEvent,
-    message: CommandMessage<typeof SCHEMA>
-  ): Promise<void> {
-    if (event !== 'CREATE') {
-      return;
-    }
-
+  async on(message: CommandMessage<typeof SCHEMA>): Promise<void> {
     const stats = await this.repo.fetchGuildStats();
     if (!stats) {
       await message.reply({
